moves channel importation logic to update channels worker.
