+++ app/build.gradle
@@ -109,11 +109,11 @@ dependencies {
     implementation 'androidx.navigation:navigation-fragment-ktx:2.3.5'
     implementation 'androidx.navigation:navigation-ui-ktx:2.3.5'
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
-    implementation "androidx.work:work-runtime:2.6.0"
     implementation "androidx.biometric:biometric:1.1.0"
     implementation 'com.google.android.material:material:1.4.0'
     implementation 'androidx.viewpager:viewpager:1.0.0'
     implementation 'androidx.recyclerview:recyclerview-selection:1.1.0'
+    implementation 'androidx.work:work-runtime-ktx:2.6.0'
     kapt "androidx.lifecycle:lifecycle-common-java8:2.4.0"
 
     implementation "androidx.room:room-runtime:2.3.0"
